requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.1
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-otp==1.3.0
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-shortuuidfield==0.1.3
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
